깊이 버퍼 추가

+++ Portfolio/GameEngineCore/GameEngineIndexBuffer.h
@@ -6,6 +6,7 @@
 #include <map>
 #include <list>
 
+// Ό³Έν :
 class GameEngineIndexBuffer : public GameEngineRes<GameEngineIndexBuffer>
 {
 	friend GameEngineRes<GameEngineIndexBuffer>;
@@ -14,16 +15,18 @@ public:
 	template<typename IndexType>
 	static GameEngineIndexBuffer* Create(const std::string& _Name, const std::vector<IndexType>& _Vertex)
 	{
-		return Create(_Name, &_Vertex[0] , sizeof(IndexType), _Vertex.size());
+		return Create(_Name, &_Vertex[0], sizeof(IndexType), static_cast<unsigned int>(_Vertex.size()));
 	}
 
 
 	static GameEngineIndexBuffer* Create(const std::string& _Name, const void* _Data, UINT _IndexSize, UINT _IndexCount);
 
 private:
+	// constrcuter destructer
 	GameEngineIndexBuffer();
 	~GameEngineIndexBuffer();
 
+	// delete Function
 	GameEngineIndexBuffer(const GameEngineIndexBuffer& _Other) = delete;
 	GameEngineIndexBuffer(GameEngineIndexBuffer&& _Other) noexcept = delete;
 	GameEngineIndexBuffer& operator=(const GameEngineIndexBuffer& _Other) = delete;
@@ -32,7 +35,7 @@ private:
 public:
 	void Setting();
 
-	inline UINT GetIndexCount() 
+	inline UINT GetIndexCount()
 	{
 		return IndexCount;
 	}
@@ -42,11 +45,17 @@ protected:
 
 
 private:
+	// nullptr
 	D3D11_BUFFER_DESC BufferDesc;
+
 	D3D11_SUBRESOURCE_DATA Data;
+
 	ID3D11Buffer* Buffer;
+
 	UINT IndexSize;
+
 	UINT IndexCount;
+
 	UINT Offset;
 };
